+++ src/App.js
@@ -1,10 +1,19 @@
-import React from 'react';
-import Main from "./components/main";
+import React from "react";
+import { Switch, Route } from "react-router-dom";
 
-function App() {
-  return (
-    <Main/>
-  );
-}
+import Login from "./login";
+import SignUp from "./signup";
+import LandingPage from "./landingpage";
+import Post from "./post";
 
-export default App;
+const Main = () => (
+  <Switch>
+        <Route exact path="/" component={Login} />
+        <Route exact path="/signup" component={SignUp} />
+        <Route exact path="/landingpage" component={LandingPage} />
+        <Route exact path="/post" component={Post} />
+    </Switch>
+)
+
+export default Main;
+
