+++ src/components/landingpage.js
@@ -1,0 +1,1 @@
+import React, { Component } from "react";
